+++ hc/accounts/forms.py
@@ -61,7 +61,7 @@ class EmailLoginForm(forms.Form):
     # to avoid some of the dumber bots
     identity = LowercaseEmailField()
 
-    def __init__(self, request: HttpRequest):
+    def __init__(self, request: HttpRequest | None = None):
         self.request = request
         super(EmailLoginForm, self).__init__(request.POST if request else None)